Support for chip types that can address more than 64kB of flash memory

+++ dumper/jtag.h
@@ -38,7 +38,7 @@ public:
 	void ping() const;
 	bool check() const;
 
-	void readFlash(uint8_t* buffer, uint16_t address, bool customBlock);
+	void readFlash(uint8_t* buffer, uint32_t address, bool customBlock);
 
 private:
 	void sendMode(uint8_t value) const;
